+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
-## 0.1.1
+## 0.1.1 ~ 0.1.2
+
+- fix: package main entry
+- fix: store undefined data
+
+## 0.1.0
 
 - initial release